BIG SMALL UPDATE. BUGFIXES. REMOVING SOME VISUAL STUDIO WARNINGS.
FIX BUFFER OVERRUN!! (Preserve crc32 option)
Update Preserve CRC option, now without bugs, and no duplicate models.

+++ src/bsp/Entity.cpp
@@ -429,8 +429,8 @@ void Entity::renameTargetnameValues(std::string oldTargetname, std::string newTa
 	}
 }
 
-int Entity::getMemoryUsage() {
-	int size = sizeof(Entity);
+size_t Entity::getMemoryUsage() {
+	size_t size = sizeof(Entity);
 
 	for (int i = 0; i < cachedTargets.size(); i++) {
 		size += cachedTargets[i].size();
