+++ src/bsp/bsptypes.cpp
@@ -4,26 +4,26 @@
 
 BSPEDGE::BSPEDGE() = default;
 
-BSPEDGE::BSPEDGE(uint16_t v1, uint16_t v2) {
+BSPEDGE::BSPEDGE(unsigned short v1, unsigned short v2) {
 	iVertex[0] = v1;
 	iVertex[1] = v2;
 }
 
 bool BSPPLANE::update(vec3 newNormal, float fdist) {
-	float fx = fabs(newNormal.x);
-	float fy = fabs(newNormal.y);
-	float fz = fabs(newNormal.z);
+	double fx = fabs(newNormal.x);
+	double fy = fabs(newNormal.y);
+	double fz = fabs(newNormal.z);
 	int planeType = PLANE_ANYZ;
 	bool shouldFlip = false;
-	if (fx > 0.9999f) {
+	if (fx > 0.9999) {
 		planeType = PLANE_X;
 		if (newNormal.x < 0) shouldFlip = true;
 	}
-	else if (fy > 0.9999f) {
+	else if (fy > 0.9999) {
 		planeType = PLANE_Y;
 		if (newNormal.y < 0) shouldFlip = true;
 	}
-	else if (fz > 0.9999f) {
+	else if (fz > 0.9999) {
 		planeType = PLANE_Z;
 		if (newNormal.z < 0) shouldFlip = true;
 	}
